Modularize sub.
