+++ tfjs-core/src/public/chained_ops/sub.ts
@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright 2020 Google Inc. All Rights Reserved.
+ * Copyright 2020 Google LLC. All Rights Reserved.
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
@@ -14,14 +14,17 @@
  * limitations under the License.
  * =============================================================================
  */
-import './add';
-import './batchnorm';
-import './broadcast_to';
-import './div';
-import './div_no_nan';
-import './one_hot';
-import './pad';
-import './squared_difference';
-import './sub';
-import './tile';
-import './transpose';
+import {sub} from '../../ops/sub';
+import {Tensor} from '../../tensor';
+import {Rank, TensorLike} from '../../types';
+
+declare module '../../tensor' {
+  interface Tensor<R extends Rank = Rank> {
+    sub<T extends Tensor>(b: Tensor|TensorLike): T;
+  }
+}
+
+Tensor.prototype.sub = function<T extends Tensor>(b: Tensor|TensorLike): T {
+  this.throwIfDisposed();
+  return sub(this, b);
+};
