+++ tfjs-core/src/tensor.ts
@@ -210,7 +210,6 @@ export interface OpHandler {
   argMax<T extends Tensor>(x: Tensor, axis: number): T;
   addStrict<T extends Tensor>(a: T, b: T|TensorLike): T;
   atan2<T extends Tensor>(a: Tensor, b: Tensor|TensorLike): T;
-  sub<T extends Tensor>(a: Tensor, b: Tensor|TensorLike): T;
   subStrict<T extends Tensor>(a: T, b: T|TensorLike): T;
   pow<T extends Tensor>(base: T, exp: Tensor|TensorLike): T;
   powStrict<T extends Tensor>(base: T, exp: Tensor|TensorLike): T;
@@ -644,7 +643,7 @@ export class Tensor<R extends Rank = Rank> {
     return this.isDisposedInternal;
   }
 
-  private throwIfDisposed() {
+  throwIfDisposed() {
     if (this.isDisposed) {
       throw new Error(`Tensor is disposed.`);
     }
@@ -883,10 +882,6 @@ export class Tensor<R extends Rank = Rank> {
   atan2<T extends this>(this: T, x: T|TensorLike): T {
     this.throwIfDisposed();
     return opHandler.atan2(this, x);
-  }
-  sub<T extends Tensor>(x: Tensor|TensorLike): T {
-    this.throwIfDisposed();
-    return opHandler.sub(this, x);
   }
   subStrict<T extends this>(this: T, x: T|TensorLike): T {
     this.throwIfDisposed();
