+++ tfjs-core/src/tests.ts
@@ -98,6 +98,7 @@ import './ops/softmax_test';
 import './ops/sparse_to_dense_test';
 import './ops/spectral_ops_test';
 import './ops/strided_slice_test';
+import './ops/sub_test';
 import './ops/tile_test';
 import './ops/topk_test';
 import './ops/transpose_test';
